edit notification banner with new content, change page it opens on, add/remove tests
