+++ packages/gafl-webapp-service/src/locales/en.json
@@ -454,6 +454,9 @@
   "newsletter_error_set_email": "Enter an email address in the correct format, like name@example.com",
   "newsletter_subscribe": "Stay up to date with the latest news on angling, fisheries and how we spend your licence money. You can unsubscribe at any time.",
   "newsletter_title": "Would you like to receive our email newsletter?",
+  "notification_heading": "Plastic-coated card licences will be replaced by a printed letter from 31 May 2024.",
+  "notification_label": "Important",
+  "notification_text": "Choosing to get a licence by email or text message helps reduce the paper we use and invests more licence income back into fishing.",
   "no_licence_req_body": "How much will this cost?",
   "no_licence_req_title": "You do not need a licence yet",
   "no": "No",
@@ -498,10 +501,6 @@
   "period": ".",
   "phone_number": "phone number",
   "pound": "£",
-  "price_increase_notification_heading": "Licence prices are changing",
-  "price_increase_notification_label": "Important",
-  "price_increase_notification_link_text": "See licence price increases (opens in new tab)",
-  "price_increase_notification_text": "Rod fishing licences starting on or after 1 April 2023 are increasing in price.",
   "pricing_summary_heading": "You can go fishing without a rod licence because you’re under 13. You’ll need to get a licence to fish as soon as you turn 13.",
   "privacy_dp_officer_body_1": "Our Data Protection Officer (DPO) is responsible for independent advice and monitoring of the Environment Agency’s use of personal data. You can read our ",
   "privacy_dp_officer_body_2": " which explains how we deal with personal data in general. If you have any concerns or queries about how we process personal data, or if you would like to make a complaint or request relating to data protection, please contact our DPO using the following details:",
