+++ packages/gafl-webapp-service/src/pages/contact/digital-licence/licence-fulfilment/__tests__/route.spec.js
@@ -41,5 +41,17 @@ describe('licence-fulfilment > route', () => {
       const result = await getData(getMockRequest())
       expect(result.youOrOther).toEqual(returnedValue)
     })
+
+    it.each([
+      ['true', true],
+      ['false', false],
+      [undefined, false]
+    ])('SHOW_NOTIFICATION_BANNER is set to value of process.env.SHOW_NOTIFICATION_BANNER', async (notification, expectedResult) => {
+      process.env.SHOW_NOTIFICATION_BANNER = notification
+      const mockRequest = getMockRequest()
+      const result = await getData(mockRequest)
+
+      expect(result.SHOW_NOTIFICATION_BANNER).toEqual(expectedResult)
+    })
   })
 })
