+++ packages/gafl-webapp-service/src/pages/contact/digital-licence/licence-fulfilment/licence-fulfilment.njk
@@ -1,5 +1,6 @@
 {% extends "standard-form.njk" %}
 {% from "radios/macro.njk" import govukRadios %}
+{% from "notification-banner.njk" import notificationBanner %}
 
 {% set title = mssgs['licence_fulfilment_title_' + data.youOrOther] %}
 
@@ -9,6 +10,12 @@
         }
     }
 %}
+
+{% block notificationContent %}
+    {% if data.SHOW_NOTIFICATION_BANNER %}
+        {{ notificationBanner(mssgs) }}
+    {% endif%}
+{% endblock %}
 
 {% block pageContent %}
     <p class="govuk-body">{{ mssgs.licence_fulfilment_content_info }}</p>
