+++ packages/gafl-webapp-service/src/pages/licence-details/licence-to-start/__tests__/route.spec.js
@@ -1,6 +1,4 @@
 import { getData } from '../route'
-import { NEW_PRICES } from '../../../../uri.js'
-import { addLanguageCodeToUri } from '../../../../processors/uri-helper.js'
 
 jest.mock('../../../../processors/uri-helper.js')
 
@@ -28,23 +26,6 @@ describe('licence-to-start > route', () => {
       const request = getMockRequest(false)
       const result = await getData(request)
       expect(result.isLicenceForYou).toBeFalsy()
-    })
-
-    it.each([
-      ['true', true],
-      ['false', false],
-      [undefined, false]
-    ])('SHOW_NOTIFICATION_BANNER is set to value of process.env.SHOW_NOTIFICATION_BANNER', async (notification, expectedResult) => {
-      process.env.SHOW_NOTIFICATION_BANNER = notification
-      const request = getMockRequest()
-      const result = await getData(request)
-      expect(result.SHOW_NOTIFICATION_BANNER).toEqual(expectedResult)
-    })
-
-    it('addLanguageCodeToUri is called with request and NEW_PRICES.uri', async () => {
-      const request = getMockRequest()
-      await getData(request)
-      expect(addLanguageCodeToUri).toHaveBeenCalledWith(request, NEW_PRICES.uri)
     })
   })
 })
