+++ packages/gafl-webapp-service/src/pages/summary/licence-summary/licence-summary.njk
@@ -4,7 +4,6 @@
 {% from "fieldset/macro.njk" import govukFieldset %}
 {% from "warning-text/macro.njk" import govukWarningText %}
 {% from "button/macro.njk" import govukButton %}
-{% from "notification-banner.njk" import notificationBanner %}
 
 {% set title = mssgs.licence_summary_title %}
 {% block pageTitle %}{{ title }}{{ mssgs.gov_uk }}{% endblock %}
@@ -27,11 +26,6 @@
 {% block content %}
 <div class="govuk-grid-row">
     <div class="govuk-grid-column-two-thirds">
-        {% block notificationContent %}
-          {% if data.SHOW_NOTIFICATION_BANNER %}
-            {{ notificationBanner(data.uri.newPrices, mssgs) }}
-          {% endif%}
-        {% endblock %}
         <form method="post">
             {{ licenseDetailsSummary | trim | safe }}
